Refactor event controller in backend
