+++ backend/src/index.ts
@@ -7,7 +7,7 @@ import groupsRouter from "../controllers/groups"
 import { unknownEndpoint, errorHandler } from "../utils/middleware"
 import routesRouter from "./routes/routes.route"
 import penaltyRouter from "./routes/penalties.route"
-import eventRouter from "../controllers/event"
+import eventRouter from "./routes/event.route"
 
 const prisma = new PrismaClient()
